+++ WebApplication3/Controllers/ProducersController.cs
@@ -13,11 +13,13 @@ namespace WebApplication3.Controllers
     public class ProducersController : Controller
     {
         private readonly IProducersService _service; //calling appdbcontext to access the context
+        private readonly IMoviesService _serviceMovie;
 
         //to access the context a constructor is required
-        public ProducersController(IProducersService service)
+        public ProducersController(IProducersService service, IMoviesService serviceMovie)
         {
             _service = service;
+            _serviceMovie = serviceMovie;
         }
 
         public async Task<IActionResult> Index()
@@ -97,19 +99,21 @@ namespace WebApplication3.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var producerDetails = await _service.GetByIdAsync(id);
+            var allMovies = await _serviceMovie.GetAllAsync(n => n.Cinema); //Ordering movies by their names          
+            var filterResult = allMovies.Where(n => n.ProducerId.Equals(id)).ToList();
 
             if (producerDetails == null)
             {
                 return View("Not Found");
             }
 
+            if (filterResult != null)
+            {
+                return View("AssignedProducers");
+            }
+
             await _service.DeleteAsync(id);
             return RedirectToAction(nameof(Index));
-
         }
-
-
-
-
     }
 }
